+++ dart/test/scope_test.dart
@@ -366,7 +366,7 @@ void main() {
       username: 'first-user',
       email: 'first@user.lan',
       ipAddress: '127.0.0.1',
-      extras: const <String, String>{'first-sign-in': '2020-01-01'},
+      data: const <String, String>{'first-sign-in': '2020-01-01'},
     );
 
     final breadcrumb = Breadcrumb(message: 'Authenticated');