+++ flutter/lib/src/integrations/on_error_integration.dart
@@ -38,10 +38,6 @@ class OnErrorIntegration implements Integration<SentryFlutterOptions> {
     // WidgetsBinding works with WidgetsFlutterBinding and other custom bindings
     final wrapper = dispatchWrapper ??
         PlatformDispatcherWrapper(binding.platformDispatcher);
-
-    if (!wrapper.isOnErrorSupported(options)) {
-      return;
-    }
     _defaultOnError = wrapper.onError;
 
     _integrationOnError = (Object exception, StackTrace stackTrace) {